Fully Featured
The addition of a settings menu, saving and loading settings to a file, change tracking, and tons of confirmation menus brings us to a more or less complete state.
It hasn't been thoroughly bugtested, but it performs as expected under not-so-rigorous conditions.
Next release will come when the byte location file is more filled out.